Fix bugs: exp() numerical precision, and loop bounds bug
I found a few small bugs while looking at @zicklag's comments on https://github.com/EmbarkStudios/texture-synthesis/pull/14

First: there is a numerical precision bug with the calculation of distance gaussians.  The exp() function used to be f64::exp(), and I was using f32::exp().

Second: there were missing entries in the precomputed function table.  Loop bounds are exclusive... but 256u8 is not a u8... so it needs 0..=255u8 which was made for this situation.