Various changes (#28)
* Inline a couple of functions

* Add pr-14 baseline

* Update CHANGELOG

* A bit more performance improvement - taking advantage of ZipIter's specialization for ExactSizeIterator

Rust iterators have some great specialization features - and .zip() has fantastic performance when joining two `DoubleEndedIterator + ExactSizeIterator` - which std::Slice::Iter implements!  The trick is that it can combine the bounds checks for the zipped-together items.  And the benches look faster than they used to be - so llvm may be optimizing the bound checks away.

These changes improve the example01 performance down to
13.06 real
39.56 user

from:
15.50 real
50.72 sys

* Cleanup CLI examples

The use of `--` is only needed when `--guides`
is passed to delimit between guide and example
paths.

Closes: #20

* Just use slices

* Remove unneeded module prefix

* Remove dependency on modulo

* Minor cleanup

* Use ImageBuffer wrapper

* Rename --out-fmt to --stdout-fmt

Resolves: #27

* Add helper func

* Add baseline for PR-28

* Update CHANGELOG

* [ci] Build and run tests in release