Change inpaint_example to require a size
Due to how inpaint works, it makes sense
to require a size when specifing an inpaint
example. There is a potential for confusion
if a user explicitly specifies resize_input
or output_size as they are ignore now if
inpaint is used, but this is noted in the
documentation, and in any case, would
previously emit a failure if you failed to set
resize_input anyways.